Refactor databases to support further implementations
